requirements.txt
aiogram==3.2.0
asyncio==3.4.3
mysql-connector-python==8.2.0
Pillow==10.1.0
request==2.31.0
datetime==5.4